refactor(api): move lcms route and controller to src
