+++ api/src/shared/application/lcms/lcms-controller.js
@@ -1,5 +1,5 @@
-import { sharedUsecases as usecases } from '../../../src/shared/domain/usecases/index.js';
-import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
+import { sharedUsecases as usecases } from '../../domain/usecases/index.js';
+import { logger } from '../../infrastructure/utils/logger.js';
 
 const createRelease = async function (request, h) {
   usecases
